backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML/AI
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
scikit-learn==1.4.0
scipy==1.12.0

# PDF Generation
reportlab==4.0.8
PyPDF2==3.0.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0.post1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Redis (for caching and pub/sub)
redis==5.0.1
